pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uni-agent"
version = "0.1.0"
description = "A pluggable-skills agent client built with Python."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.7,<3",
  "pydantic-settings>=2.2,<3",
  "python-dotenv>=1.0,<2",
  "PyYAML>=6.0,<7",
  "structlog>=24.1,<25",
  "typer>=0.12,<1",
  "pydantic-ai>=0.0.24,<2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
]

[project.scripts]
agent = "uni_agent.cli.main:app"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
